Remove warnings, add MCtruth attributes to KFitParticle, different masses can be used in mass fits

+++ source/KFitDecayBuilder.cxx
@@ -17,8 +17,8 @@
 #include "KFitDecayBuilder.h"
 
 KFitDecayBuilder::KFitDecayBuilder(TString task, std::vector<int> pids, TLorentzVector lv, double mass) : fTask(task),
-                                                                                                            fPids(pids),
-                                                                                                            fVerbose(0)
+                                                                                                          fPids(pids),
+                                                                                                          fVerbose(0)
 
 {
     if (fVerbose > 0)
@@ -28,7 +28,6 @@ KFitDecayBuilder::KFitDecayBuilder(TString task, std::vector<int> pids, TLorentz
 
     setIniSys(lv);
     setMass(mass);
-
 }
 
 void KFitDecayBuilder::countCombis()
@@ -40,13 +39,13 @@ void KFitDecayBuilder::countCombis()
     fCombiCounter = 0;
     // Determine the number of combinations of the input particles
     fTotalCombis = 1;
-    if(particleCounter.size()>0) particleCounter.clear();
+    if (particleCounter.size() > 0)
+        particleCounter.clear();
     for (size_t i = 0; i < fPids.size(); i++)
     {
         fTotalCombis *= fCands[i].size();
         particleCounter.push_back(0);
     }
-
 }
 
 void KFitDecayBuilder::buildDecay()
@@ -58,18 +57,18 @@ void KFitDecayBuilder::buildDecay()
     if (fVerbose > 1)
     {
         std::cout << "Combi counter: " << fCombiCounter << " total Combos: " << fTotalCombis << std::endl;
-        std::cout<< "Task is " << fTask <<std::endl;
+        std::cout << "Task is " << fTask << std::endl;
     }
 
     // For selection of best event according to probability
     fBestProb = 0;
     fBestChi2 = 1e6;
 
-    while (fCombiCounter < (fTotalCombis ))
+    while (fCombiCounter < (fTotalCombis))
     {
         // Take one combination of particles
         fillFitCands();
-        
+
         // Check if correct number of particles
         if (fFitCands.size() != fPids.size())
         {
@@ -91,13 +90,9 @@ void KFitDecayBuilder::buildDecay()
         if (fVerbose > 1)
         {
             std::cout << "Combi counter: " << fCombiCounter << " total Combos: " << fTotalCombis << std::endl;
-            std::cout << "fit finished" <<std::endl;
+            std::cout << "fit finished" << std::endl;
         }
-        
-        
-        
     }
-    
 }
 
 Bool_t KFitDecayBuilder::doFit()
@@ -107,12 +102,18 @@ Bool_t KFitDecayBuilder::doFit()
         std::cout << "--------------- KFitDecayBuilder::doFit() -----------------" << std::endl;
     }
     KinFitter Fitter(fFitCands);
-    if (fTask == "4C") Fitter.add4Constraint(fIniSys);
-    else if (fTask == "Vertex") Fitter.addVertexConstraint();
-    else if (fTask == "MassVtx") Fitter.addMassVtxConstraint(fMass);
-    else if (fTask == "MM" || fTask == "MissingMass" ) Fitter.addMissingMassConstraint(fIniSys, fMass);
-    else if (fTask == "Mom" || fTask == "MissingParticle") Fitter.addMissingParticleConstraint(fIniSys, fMass);
-    else if (fTask == "Mass"){
+    if (fTask == "4C")
+        Fitter.add4Constraint(fIniSys);
+    else if (fTask == "Vertex")
+        Fitter.addVertexConstraint();
+    else if (fTask == "MassVtx")
+        Fitter.addMassVtxConstraint(fMass);
+    else if (fTask == "MM" || fTask == "MissingMass")
+        Fitter.addMissingMassConstraint(fIniSys, fMass);
+    else if (fTask == "Mom" || fTask == "MissingParticle")
+        Fitter.addMissingParticleConstraint(fIniSys, fMass);
+    else if (fTask == "Mass")
+    {
         Fitter.addMassConstraint(fMass);
         /*
         cout << "constraint added, Mass = " << fMass << endl;
@@ -120,12 +121,12 @@ Bool_t KFitDecayBuilder::doFit()
         cout << "pion momentum = " << fFitCands[1].getMomentum() << endl;
         TMatrixD cov_p = fFitCands[0].getCovariance();
         TMatrixD cov_pi = fFitCands[1].getCovariance();
-        /*
         cout << "proton momentum error = " << cov_p(0,0) << endl;
         cout << "pion momentum error = " << cov_pi(0,0) << endl;
-*/
-    } 
-    else cout << "Task not available" << endl;
+        */
+    }
+    else
+        cout << "Task not available" << endl;
 
     if (Fitter.fit())
     {
@@ -137,7 +138,8 @@ Bool_t KFitDecayBuilder::doFit()
         {
             fBestProb = Fitter.getProb();
             fBestChi2 = Fitter.getChi2();
-            if(fOutputCands.size() != 0) fOutputCands.clear();
+            if (fOutputCands.size() != 0)
+                fOutputCands.clear();
             Fitter.getDaughters(fOutputCands);
         }
         return kTRUE;
@@ -159,9 +161,10 @@ void KFitDecayBuilder::fillFitCands()
         std::cout << "--------------- KFitDecayBuilder::fillFitCands() -----------------" << std::endl;
     }
 
-    if(fFitCands.size()>0) fFitCands.clear();
+    if (fFitCands.size() > 0)
+        fFitCands.clear();
     doubleParticle = false;
-    
+
     for (size_t i = 0; i < fPids.size(); i++)
     {
         checkDoubleParticle(i);
@@ -174,7 +177,7 @@ void KFitDecayBuilder::fillFitCands()
     int a = fPids.size() - 1;
     if (fVerbose > 1)
         std::cout << "particle counter " << a << " is " << particleCounter[a] << " now" << std::endl;
-    while (particleCounter[a] == (fCands[a].size() - 1))
+    while (particleCounter[a] == (((int)fCands[a].size()) - 1))
     {
         particleCounter[a] = 0;
         if (fVerbose > 1)
@@ -182,9 +185,9 @@ void KFitDecayBuilder::fillFitCands()
         a--;
         if (a < 0)
         {
-            if (!(fCombiCounter == fTotalCombis-1))
-            if (fVerbose > 1)
-                std::cout << "counted wrong: " << fCombiCounter << " != " << fTotalCombis << std::endl;
+            if (!(fCombiCounter == fTotalCombis - 1))
+                if (fVerbose > 1)
+                    std::cout << "counted wrong: " << fCombiCounter << " != " << fTotalCombis << std::endl;
             break;
 
             a = 1e6;
@@ -194,7 +197,6 @@ void KFitDecayBuilder::fillFitCands()
     if (fVerbose > 1)
         std::cout << "particle counter " << a << " is " << particleCounter[a] << " now" << std::endl;
     fCombiCounter++;
-    
 
     if (doubleParticle && (fCombiCounter < fTotalCombis))
         fillFitCands(); // If some particle has been filled more than once into fFitCands, repeat the procedure with the next combination
